chore(refactor): moves node UI components to global location
