+++ frontend/packages/syntara-ui/src/components/nodes/NodeComponent.tsx
@@ -4,15 +4,15 @@ import { ExecutorTypeEnum, type TaskActivity } from '@syntara/contracts'
 import { Handle, type NodeProps, Position } from '@xyflow/react'
 import React, { useEffect, useMemo, useState } from 'react'
 
-import { SynPanel } from '../../../../../components/layout/SynPanel'
-import { FlowNodeType } from '../../../../../constants'
-import { ExecutionStatusBadge } from '../../../../builder/components/ExecutionStatusBadge'
-import type { ActivityStatus } from '../../../execution/types'
-import { NODE_TYPE_COLORS } from '../../nodeTypeColors'
-import type { SemanticZoomBranchSource } from '../../semanticZoomTypes'
-import { useSemanticZoom } from '../hooks/useSemanticZoom'
+import { SynPanel } from '../layout/SynPanel'
+import { FlowNodeType } from '../../constants'
+import { ExecutionStatusBadge } from '../../routes/builder/components/ExecutionStatusBadge'
+import type { ActivityStatus } from '../../routes/workflows/execution/types'
+import { NODE_TYPE_COLORS } from '../../routes/workflows/canvas/nodeTypeColors'
+import type { SemanticZoomBranchSource } from '../../routes/workflows/canvas/semanticZoomTypes'
+import { useSemanticZoom } from '../../routes/workflows/canvas/nodes/hooks/useSemanticZoom'
 
-import { targetHandleStyle, sourceHandleStyle } from './handleStyle'
+import { targetHandleStyle, sourceHandleStyle } from '../../routes/workflows/canvas/nodes/common/handleStyle'
 import { NodeExpandedAllContext } from './NodeExpandedAllContext'
 import { NodeExpandedContext } from './NodeExpandedContext'
 import { NodeSemanticZoomBody } from './NodeSemanticZoomBody'
