+++ frontend/packages/syntara-ui/src/components/nodes/NodeExpandToggle.tsx
