+++ frontend/packages/syntara-ui/src/components/nodes/NodeExpandedAllContext.test.tsx
