+++ frontend/packages/syntara-ui/src/components/nodes/NodeExpandedAllContext.tsx
