+++ frontend/packages/syntara-ui/src/components/nodes/NodeHeader.test.tsx
