+++ frontend/packages/syntara-ui/src/components/nodes/NodeSemanticZoomBody.test.tsx
