+++ frontend/packages/syntara-ui/src/components/nodes/NodeSidePanel.test.tsx
