+++ frontend/packages/syntara-ui/src/routes/workflows/canvas/nodes/TaskNode.tsx
@@ -8,8 +8,8 @@ import { getNodeTypeColor } from '../nodeTypeColors'
 
 import { renderCondition, renderJson, renderText } from './common/detailRenderers'
 import { detectTaskNodeType, type TaskActivityWithMetadata } from './common/detectTaskNodeType'
-import { NodeBody } from './common/NodeBody'
-import { NodeComponent } from './common/NodeComponent'
+import { NodeBody } from '../../../../components/nodes/NodeBody'
+import { NodeComponent } from '../../../../components/nodes/NodeComponent'
 import { StandardNodeHeader } from './common/StandardNodeHeader'
 import { useCredentialName } from './hooks/useCredentialName'
 import { MenuNodeType, useNodeMenuActions } from './hooks/useNodeMenuActions'
