+++ frontend/packages/syntara-ui/src/routes/workflows/canvas/nodes/TaskReversedNode.tsx
@@ -5,7 +5,7 @@ import { FlowNodeType } from '../../../../constants'
 import type { ActivityStatus } from '../../execution/types'
 import { getNodeTypeColor } from '../nodeTypeColors'
 
-import { NodeComponent } from './common/NodeComponent'
+import { NodeComponent } from '../../../../components/nodes/NodeComponent'
 import { nodeMetadata } from './nodeMetadata'
 import { TaskActivityDetails } from './TaskNode'
 import { getTaskSemanticLabels } from './taskSemanticLabels'
